Unit tests added
Unit tests have been added as well as improvements in service logging and instrumentation

+++ backend/pkg/enroller/api/service.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"context"
 	"crypto/x509"
+	"database/sql"
 	"enroller/pkg/enroller/auth"
 	"enroller/pkg/enroller/crypto"
 	"enroller/pkg/enroller/models/certs"
@@ -14,6 +15,7 @@ import (
 	"enroller/pkg/enroller/secrets"
 	"errors"
 	"fmt"
+	"os"
 	"strconv"
 	"strings"
 	"sync"
@@ -43,10 +45,28 @@ type enrollerService struct {
 }
 
 var (
-	ErrIncorrectType    = errors.New("Incorrect media type")
-	ErrIncorrectContent = errors.New("Incorrect data content")
-	ErrBadRouting       = errors.New("Bad routing")
-	ErrBadKey           = errors.New("Unexpected JWT key signing method")
+	// Client errors
+	ErrInvalidCSR       = errors.New("unable to parse CSR, is invalid") //400
+	ErrInvalidID        = errors.New("invalid CSR ID, does not exist")  //404
+	ErrInvalidIDFormat  = errors.New("invalid ID format")
+	ErrInvalidApprobeOp = errors.New("invalid operation, only pending status CSRs can be approved") //400
+	ErrInvalidRevokeOp  = errors.New("invalid operation, only approved status CSRs can be revoked") //400
+	ErrInvalidDenyOp    = errors.New("invalid operation, only pending status CSRs can be denied")   //400
+	ErrInvalidDeleteOp  = errors.New("invalid operation, only pending status CSRs can be deleted")  //400
+	ErrIncorrectType    = errors.New("unsupported media type")                                      //415
+	ErrEmptyBody        = errors.New("empty body")
+
+	//Server errors
+	ErrInvalidOperation = errors.New("invalid operation")
+	ErrInsertCSR        = errors.New("unable to insert CSR")
+	ErrInsertCert       = errors.New("unable to insert certificate")
+	ErrGetCSR           = errors.New("unable to get CSR")
+	ErrGetCert          = errors.New("unable to get certificate")
+	ErrUpdateCSR        = errors.New("unable to update CSR")
+	ErrDeleteCSR        = errors.New("unable to delete CSR")
+	ErrSignCSR          = errors.New("unable to sign CSR")
+	ErrRevokeCert       = errors.New("unable to revoke certificate")
+	ErrResponseEncode   = errors.New("error encoding response")
 )
 
 func NewEnrollerService(csrDBStore csrstore.DB, csrFileStore csrstore.File, certsDBStore certstore.DB, certsFileStore certstore.File, secrets secrets.Secrets, homePath string) Service {
@@ -61,7 +81,11 @@ func NewEnrollerService(csrDBStore csrstore.DB, csrFileStore csrstore.File, cert
 }
 
 func (s *enrollerService) PostCSR(ctx context.Context, data []byte) (csrmodel.CSR, error) {
-	csr, err := s.insertCSRInDB(data)
+	csr, err := parseCSRDataModel(data)
+	if err != nil {
+		return csrmodel.CSR{}, err
+	}
+	csr, err = s.insertCSRInDB(csr)
 	if err != nil {
 		return csrmodel.CSR{}, err
 	}
@@ -72,10 +96,10 @@ func (s *enrollerService) PostCSR(ctx context.Context, data []byte) (csrmodel.CS
 	return csr, nil
 }
 
-func (s *enrollerService) insertCSRInDB(data []byte) (csrmodel.CSR, error) {
+func parseCSRDataModel(data []byte) (csrmodel.CSR, error) {
 	certReq, err := crypto.ParseNewCSR(data)
 	if err != nil {
-		return csrmodel.CSR{}, ErrIncorrectContent
+		return csrmodel.CSR{}, ErrInvalidCSR
 	}
 	csr := csrmodel.CSR{
 		CountryName:            strings.Join(certReq.Subject.Country, " "),
@@ -87,15 +111,19 @@ func (s *enrollerService) insertCSRInDB(data []byte) (csrmodel.CSR, error) {
 		CommonName:             certReq.Subject.CommonName,
 		Status:                 csrmodel.PendingStatus,
 	}
+	return csr, nil
+}
+
+func (s *enrollerService) insertCSRInDB(csr csrmodel.CSR) (csrmodel.CSR, error) {
 	id, err := s.csrDBStore.Insert(csr)
 	if err != nil {
-		return csrmodel.CSR{}, ErrIncorrectContent
+		return csrmodel.CSR{}, ErrInsertCSR
 	}
 	csr.Id = id
 	csr.CsrFilePath = s.homePath + "/" + strconv.Itoa(id) + ".csr"
 	err = s.csrDBStore.UpdateFilePath(csr)
 	if err != nil {
-		return csrmodel.CSR{}, ErrIncorrectContent
+		return csrmodel.CSR{}, ErrInsertCSR
 	}
 	return csr, nil
 }
@@ -103,8 +131,8 @@ func (s *enrollerService) insertCSRInDB(data []byte) (csrmodel.CSR, error) {
 func (s *enrollerService) insertCSRFile(data []byte, id int) error {
 	err := s.csrFileStore.Insert(id, data)
 	if err != nil {
-		s.csrDBStore.UpdateByID(id, csrmodel.CSR{Status: csrmodel.DeniedStatus})
-		return ErrIncorrectContent
+		s.csrDBStore.Delete(id)
+		return ErrInsertCSR
 	}
 	return nil
 }
@@ -124,7 +152,10 @@ func (s *enrollerService) GetPendingCSRs(ctx context.Context) csrmodel.CSRs {
 func (s *enrollerService) GetPendingCSRDB(ctx context.Context, id int) (csrmodel.CSR, error) {
 	c, err := s.csrDBStore.SelectByID(id)
 	if err != nil {
-		return csrmodel.CSR{}, ErrIncorrectContent
+		if err == sql.ErrNoRows {
+			return csrmodel.CSR{}, ErrInvalidID
+		}
+		return csrmodel.CSR{}, ErrGetCSR
 	}
 	return c, nil
 }
@@ -132,30 +163,59 @@ func (s *enrollerService) GetPendingCSRDB(ctx context.Context, id int) (csrmodel
 func (s *enrollerService) GetPendingCSRFile(ctx context.Context, id int) ([]byte, error) {
 	data, err := s.csrFileStore.SelectByID(id)
 	if err != nil {
-		return nil, ErrIncorrectContent
+		if os.IsNotExist(err) {
+			return nil, ErrInvalidID
+		}
+		return nil, ErrGetCSR
 	}
 	return data, nil
 }
 
 func (s *enrollerService) PutChangeCSRStatus(ctx context.Context, csr csrmodel.CSR, id int) (csrmodel.CSR, error) {
 	var err error
+
+	prevCSR, err := s.csrDBStore.SelectByID(id)
+	if err != nil {
+		if err == sql.ErrNoRows {
+			return csrmodel.CSR{}, ErrInvalidID
+		}
+		return csrmodel.CSR{}, ErrGetCSR
+	}
+
 	switch status := csr.Status; status {
 	case csrmodel.ApprobedStatus:
-		err = s.approbeCSR(id, csr)
-	case csrmodel.RevokedStatus:
-		_, err = s.csrDBStore.UpdateByID(id, csr)
-		if err != nil {
-			return csrmodel.CSR{}, err
+		if prevCSR.Status == csrmodel.PendingStatus {
+			err = s.approbeCSR(id, csr)
+			if err != nil {
+				return csrmodel.CSR{}, err
+			}
+		} else {
+			return csrmodel.CSR{}, ErrInvalidApprobeOp
 		}
-		err = s.revokeCert(id)
-	case csrmodel.DeniedStatus, csrmodel.PendingStatus:
-		_, err = s.csrDBStore.UpdateByID(id, csr)
+	case csrmodel.RevokedStatus:
+		if prevCSR.Status == csrmodel.ApprobedStatus {
+			_, err = s.csrDBStore.UpdateByID(id, csr)
+			if err != nil {
+				return csrmodel.CSR{}, ErrUpdateCSR
+			}
+			err = s.revokeCert(id)
+			if err != nil {
+				return csrmodel.CSR{}, err
+			}
+		} else {
+			return csrmodel.CSR{}, ErrInvalidRevokeOp
+		}
+	case csrmodel.DeniedStatus:
+		if prevCSR.Status == csrmodel.PendingStatus {
+			_, err = s.csrDBStore.UpdateByID(id, csr)
+			if err != nil {
+				return csrmodel.CSR{}, ErrUpdateCSR
+			}
+		} else {
+			return csrmodel.CSR{}, ErrInvalidDenyOp
+		}
 	default:
-		err = ErrIncorrectContent
-
-	}
-	if err != nil {
-		return csrmodel.CSR{}, err
+		return csrmodel.CSR{}, ErrInvalidOperation
 	}
 
 	return csr, nil
@@ -165,7 +225,7 @@ func (s *enrollerService) revokeCert(id int) error {
 	revocationDate := makeOpenSSLTime(time.Now())
 	err := s.certsDBStore.Revoke(id, revocationDate)
 	if err != nil {
-		return ErrIncorrectContent
+		return ErrRevokeCert
 	}
 	return nil
 
@@ -186,11 +246,11 @@ func (s *enrollerService) approbeCSR(id int, csr csrmodel.CSR) error {
 	}
 	err = s.insertCertFile(id, crt)
 	if err != nil {
-		return err
+		return ErrInsertCert
 	}
 	_, err = s.csrDBStore.UpdateByID(id, csr)
 	if err != nil {
-		return ErrIncorrectContent
+		return ErrUpdateCSR
 	}
 	return nil
 
@@ -199,12 +259,12 @@ func (s *enrollerService) approbeCSR(id int, csr csrmodel.CSR) error {
 func (s *enrollerService) signCSR(csr *x509.CertificateRequest) (*x509.Certificate, error) {
 	crtData, err := s.secrets.SignCSR(csr)
 	if err != nil {
-		return nil, ErrIncorrectContent
+		return nil, ErrSignCSR
 	}
 
 	crt, err := x509.ParseCertificate(crtData)
 	if err != nil {
-		return nil, ErrIncorrectContent
+		return nil, ErrSignCSR
 	}
 	return crt, nil
 }
@@ -212,11 +272,14 @@ func (s *enrollerService) signCSR(csr *x509.CertificateRequest) (*x509.Certifica
 func (s *enrollerService) readCSRFromFile(id int) (*x509.CertificateRequest, error) {
 	csrData, err := s.csrFileStore.SelectByID(id)
 	if err != nil {
-		return nil, ErrBadRouting
+		if err == sql.ErrNoRows {
+			return nil, ErrInvalidID
+		}
+		return nil, ErrGetCSR
 	}
 	csr, err := crypto.ParseNewCSR(csrData)
 	if err != nil {
-		return nil, ErrBadRouting
+		return nil, ErrGetCSR
 	}
 	return csr, nil
 }
@@ -238,7 +301,7 @@ func (s *enrollerService) insertCertInDB(id int, crt *x509.Certificate) error {
 	}
 	err := s.certsDBStore.Insert(cert)
 	if err != nil {
-		return err
+		return ErrInsertCert
 	}
 	return nil
 }
@@ -246,29 +309,41 @@ func (s *enrollerService) insertCertInDB(id int, crt *x509.Certificate) error {
 func (s *enrollerService) insertCertFile(id int, crt *x509.Certificate) error {
 	err := s.certsFileStore.Insert(id, crt.Raw)
 	if err != nil {
-		return ErrIncorrectContent
+		return ErrInsertCert
 	}
 	return nil
 }
 
 func (s *enrollerService) DeleteCSR(ctx context.Context, id int) error {
-	_, err := s.csrDBStore.UpdateByID(id, csr.CSR{Status: csr.DeniedStatus})
+	csr, err := s.csrDBStore.SelectByID(id)
 	if err != nil {
-		return ErrIncorrectContent
+		if err == sql.ErrNoRows {
+			return ErrInvalidID
+		}
+		return ErrGetCSR
 	}
-	err = s.csrFileStore.Delete(id)
-	if err != nil {
-		return ErrIncorrectContent
+	if csr.Status == "NEW" {
+		err = s.csrDBStore.Delete(id)
+		if err != nil {
+			return ErrDeleteCSR
+		}
+		err = s.csrFileStore.Delete(id)
+		if err != nil {
+			return ErrDeleteCSR
+		}
+		return nil
 	}
-	return nil
+	return ErrInvalidDeleteOp
 }
 
 func (s *enrollerService) GetCRT(ctx context.Context, id int) ([]byte, error) {
 	data, err := s.certsFileStore.SelectByID(id)
 	if err != nil {
-		return nil, ErrBadRouting
+		if os.IsNotExist(err) {
+			return nil, ErrInvalidID
+		}
+		return nil, ErrGetCert
 	}
-
 	return data, nil
 }
